requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
idna==3.4
oauthlib==3.2.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.1
sqlparse==0.4.4
tweepy==4.14.0
tzdata==2023.3
urllib3==2.0.7
